+++ include/ignition/math/Matrix3.hh
@@ -610,15 +610,16 @@ namespace ignition
       public: friend std::ostream &operator<<(
                   std::ostream &_out, const ignition::math::Matrix3<T> &_m)
       {
-        _out << precision(_m(0, 0), 6) << " "
-             << precision(_m(0, 1), 6) << " "
-             << precision(_m(0, 2), 6) << " "
-             << precision(_m(1, 0), 6) << " "
-             << precision(_m(1, 1), 6) << " "
-             << precision(_m(1, 2), 6) << " "
-             << precision(_m(2, 0), 6) << " "
-             << precision(_m(2, 1), 6) << " "
-             << precision(_m(2, 2), 6);
+        for (auto i : {0, 1, 2})
+        {
+          for (auto j : {0, 1, 2})
+          {
+            if (!(i == 0 && j == 0))
+              _out << " ";
+
+            appendToStream(_out, _m(i, j), 6);
+          }
+        }
 
         return _out;
       }