+++ include/ignition/math/Vector3.hh
@@ -741,8 +741,14 @@ namespace ignition
       public: friend std::ostream &operator<<(
                   std::ostream &_out, const ignition::math::Vector3<T> &_pt)
       {
-        _out << precision(_pt[0], 6) << " " << precision(_pt[1], 6) << " "
-          << precision(_pt[2], 6);
+        for (auto i : {0, 1, 2})
+        {
+          if (i > 0)
+            _out << " ";
+
+          appendToStream(_out, _pt[i], 6);
+        }
+
         return _out;
       }