+++ src/Color_TEST.cc
@@ -330,10 +330,10 @@ TEST(Color, ConstAndSet)
 /////////////////////////////////////////////////
 TEST(Color, OperatorStreamOut)
 {
-  math::Color c(0.1f, 0.2f, 0.3f, 0.5f);
+  math::Color c(0.1f, 0.2f, 0.3f, 0.0f);
   std::ostringstream stream;
   stream << c;
-  EXPECT_EQ(stream.str(), "0.1 0.2 0.3 0.5");
+  EXPECT_EQ(stream.str(), "0.1 0.2 0.3 0");
 }
 
 /////////////////////////////////////////////////