+++ src/Matrix4_TEST.cc
@@ -548,11 +548,11 @@ TEST(Matrix4dTest, OperatorStreamOut)
   math::Matrix4d matA(1, 2, 3, 4,
                       5, 6, 7, 8,
                       9, 10, 11, 12,
-                      13, 14, 15, 16);
+                      13, 14, 15, 0);
 
   std::ostringstream stream;
   stream << matA;
-  EXPECT_EQ(stream.str(), "1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16");
+  EXPECT_EQ(stream.str(), "1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 0");
 }
 
 /////////////////////////////////////////////////